refactor(user): UserQueryRepository moved to query package.
