+++ src/main/java/br/com/fiapbook/user/infrastructure/repository/UserRepository.java
@@ -1,5 +1,6 @@
 package br.com.fiapbook.user.infrastructure.repository;
 
+import br.com.fiapbook.user.infrastructure.repository.query.UserQueryRepository;
 import br.com.fiapbook.user.model.entity.User;
 import java.util.Optional;
 import java.util.UUID;
@@ -7,7 +8,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<User, UUID>, UserQueryRepository{
+public interface UserRepository extends JpaRepository<User, UUID>, UserQueryRepository {
 
   Optional<User> findByEmail(String email);
